review travis scripts

+++ scripts/install/install_linux_generate_and_build.sh
@@ -1,27 +1,14 @@
-if [ ! -d "/usr/local/bin" ]; then
-  sudo mkdir -p /usr/local/bin
-fi
-
-if [ ! -d "/usr/local/cmake" ]; then
-  sudo mkdir -p /usr/local/cmake
-fi
-
-if [ ! -d "/usr/local/include" ]; then
-  sudo mkdir -p /usr/local/include
-fi
-
-if [ ! -d "/usr/local/lib" ]; then
-  sudo mkdir -p /usr/local/lib
-fi
+if [ ! -d "/usr/local/bin" ]; then sudo mkdir -p /usr/local/bin; fi
+if [ ! -d "/usr/local/cmake" ]; then sudo mkdir -p /usr/local/cmake; fi
+if [ ! -d "/usr/local/include" ]; then sudo mkdir -p /usr/local/include; fi
+if [ ! -d "/usr/local/lib" ]; then sudo mkdir -p /usr/local/lib; fi
 
 sudo chown -R $(whoami) /usr/local/bin
 sudo chown -R $(whoami) /usr/local/cmake
 sudo chown -R $(whoami) /usr/local/include
 sudo chown -R $(whoami) /usr/local/lib
 
-if [ -d ./build ]; then
-  rm -r -f build
-fi
+if [ -d ./build ]; then rm -r -f build; fi
 
 mkdir -p build/3rdparty
 cd build/3rdparty
